requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
psycopg-pool==3.2.4
sqlparse==0.5.3
langgraph==0.2.60
langchain-openai==0.2.14
langchain-community==0.3.14
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.10.1